Incorporate Francesco's suggestions, clean up code a little
