+++ plugins/formal-verification/formver.core/src/org/jetbrains/kotlin/formver/embeddings/LambdaExp.kt
@@ -6,12 +6,14 @@
 package org.jetbrains.kotlin.formver.embeddings
 
 import org.jetbrains.kotlin.fir.declarations.FirAnonymousFunction
-import org.jetbrains.kotlin.formver.conversion.*
+import org.jetbrains.kotlin.formver.conversion.MethodConversionContext
+import org.jetbrains.kotlin.formver.conversion.ResultTrackingContext
+import org.jetbrains.kotlin.formver.conversion.StmtConversionContext
+import org.jetbrains.kotlin.formver.conversion.insertInlineFunctionCall
 import org.jetbrains.kotlin.formver.embeddings.callables.CallableEmbedding
 import org.jetbrains.kotlin.formver.embeddings.callables.FunctionSignature
 import org.jetbrains.kotlin.formver.embeddings.callables.asData
 import org.jetbrains.kotlin.formver.viper.ast.Exp
-import org.jetbrains.kotlin.formver.viper.ast.Stmt
 
 class LambdaExp(
     val signature: FunctionSignature,
@@ -23,23 +25,9 @@ class LambdaExp(
 
     override fun toViper(): Exp = TODO("create new function object with counter, duplicable (requires toViper restructuring)")
 
-    override fun insertCallImpl(args: List<ExpEmbedding>, ctx: StmtConversionContext<ResultTrackingContext>): ExpEmbedding =
-        ctx.withResult(returnType) {
-            val inlineBody = function.body ?: throw Exception("Lambda $function has a null body")
-            val paramNames = function.valueParameters.map { it.name }
-            val callArgs = ctx.getInlineFunctionCallArgs(args)
-            val subs = paramNames.zip(callArgs).toMap()
-            val returnLabelName = ReturnLabelName(newWhileIndex())
-            val newCtx = MethodConverter(
-                this, this@LambdaExp.signature,
-                InlineParameterResolver(this.resultCtx.resultVar.name, returnLabelName, subs),
-                parentCtx
-            )
-            val lambdaCtx = this.newBlock().withMethodContext(newCtx)
-            lambdaCtx.convert(inlineBody)
-            lambdaCtx.addDeclaration(lambdaCtx.returnLabel.toDecl())
-            lambdaCtx.addStatement(lambdaCtx.returnLabel.toStmt())
-            // NOTE: Putting the block inside the then branch of an if-true statement is a little hack to make Viper respect the scoping
-            addStatement(Stmt.If(Exp.BoolLit(true), lambdaCtx.block, Stmt.Seqn(listOf(), listOf())))
-        }
+    override fun insertCallImpl(args: List<ExpEmbedding>, ctx: StmtConversionContext<ResultTrackingContext>): ExpEmbedding {
+        val inlineBody = function.body ?: throw Exception("Lambda $function has a null body")
+        val paramNames = function.valueParameters.map { it.name }
+        return ctx.insertInlineFunctionCall(signature, paramNames, args, inlineBody, parentCtx)
+    }
 }
