+++ plugins/formal-verification/formver.core/src/org/jetbrains/kotlin/formver/embeddings/callables/InlineNamedFunction.kt
@@ -7,34 +7,19 @@ package org.jetbrains.kotlin.formver.embeddings.callables
 
 import org.jetbrains.kotlin.fir.symbols.SymbolInternals
 import org.jetbrains.kotlin.fir.symbols.impl.FirFunctionSymbol
-import org.jetbrains.kotlin.formver.conversion.*
+import org.jetbrains.kotlin.formver.conversion.ResultTrackingContext
+import org.jetbrains.kotlin.formver.conversion.StmtConversionContext
+import org.jetbrains.kotlin.formver.conversion.insertInlineFunctionCall
 import org.jetbrains.kotlin.formver.embeddings.ExpEmbedding
-import org.jetbrains.kotlin.formver.viper.ast.Exp
-import org.jetbrains.kotlin.formver.viper.ast.Stmt
 
 class InlineNamedFunction(
     val signature: FullNamedFunctionSignature,
     val symbol: FirFunctionSymbol<*>,
 ) : CallableEmbedding, FullNamedFunctionSignature by signature {
     @OptIn(SymbolInternals::class)
-    override fun insertCallImpl(args: List<ExpEmbedding>, ctx: StmtConversionContext<ResultTrackingContext>): ExpEmbedding =
-        ctx.withResult(returnType) {
-            val inlineBody = symbol.fir.body ?: throw Exception("Function symbol $symbol has a null body")
-            val paramNames = symbol.valueParameterSymbols.map { it.name }
-            val callArgs = getInlineFunctionCallArgs(args)
-            val subs = paramNames.zip(callArgs).toMap()
-            val returnLabelName = ReturnLabelName(ctx.newWhileIndex())
-            val newCtx = MethodConverter(
-                this,
-                this@InlineNamedFunction.signature,
-                InlineParameterResolver(resultCtx.resultVar.name, returnLabelName, subs)
-            )
-            val inlineCtx = newBlock().withMethodContext(newCtx)
-            inlineCtx.convert(inlineBody)
-            // TODO: add these labels automatically.
-            inlineCtx.addDeclaration(inlineCtx.returnLabel.toDecl())
-            inlineCtx.addStatement(inlineCtx.returnLabel.toStmt())
-            // Note: Putting the block inside the then branch of an if-true statement is a little hack to make Viper respect the scoping
-            addStatement(Stmt.If(Exp.BoolLit(true), inlineCtx.block, Stmt.Seqn(listOf(), listOf())))
-        }
+    override fun insertCallImpl(args: List<ExpEmbedding>, ctx: StmtConversionContext<ResultTrackingContext>): ExpEmbedding {
+        val inlineBody = symbol.fir.body ?: throw Exception("Function symbol $symbol has a null body")
+        val paramNames = symbol.valueParameterSymbols.map { it.name }
+        return ctx.insertInlineFunctionCall(signature, paramNames, args, inlineBody)
+    }
 }
